requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
pandas>=2.0.0
plotly>=5.18.0
gspread>=5.12.0
oauth2client>=4.1.3
authlib>=1.2.0
requests>=2.31.0
bcrypt>=4.1.0
